Overhaul shared folders
Adds granularity for the shared folders so operation with the
vagrant-fsnotify plugin is more useful. Even though you can exclude
folders with fsnotify this just means the notifications are exluded -
the underlying listen gem still walks the files.

The default is the same as it was but an environment can now stipulate
the granular folder sharing and the notification excludes in the env yaml.